Cleaned up handlers
Categorized handlers into separate files
